+++ R/print.R
@@ -15,7 +15,7 @@ print.BirdFlow <- function(x, ...) {
   cat("BirdFlow model\n")
 
   nc <- 15 # n characters before colon
-  pad <- function(x) stringr::str_pad(x, width = nc, side = "right", pad = " ")
+  pad <- function(x) format(x, width = nc)
   cat(pad("  dimensions"), ": ",
       paste(nrow(x), ncol(x), n_timesteps(x), sep = ", "),
       "  (nrow, ncol, ntimesteps)\n", sep = "")